client, receiving string char by char, need to fix server

+++ client.h
@@ -29,7 +29,7 @@ char *inputMessage(FILE* fp, size_t size);
 
 bool receivedHandshake(int s);
 
-void getStringFromRecv(int s, char * str, uint8_t len);
+void getStringFromRecv(int s, char * str, int len);
 void receiveMessage(int s, uint8_t flag, struct username * users, uint16_t* numberOfUsers); // expecting length string (msglen msg is flag is 0x00)
 
 void addUserName(struct username * users, uint16_t* numberOfUsers, char* name, int len);
